LayoutTests/media/modern-media-controls/tracks-support/auto-text-track-expected.txt

Ensure we have the Auto text track selected by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
Setting caption display mode to 'automatic'...
Tapping tracks button...
PASS checkedItems.length is 1
PASS checkedItems[0].title is "Auto (Recommended)"
Setting caption display mode to 'forcedonly'...
PASS successfullyParsed is true

TEST COMPLETE
